[smooth]: Moved code from library module to library functions
