+++ src/subpackages/library/cellprofiler_library/functions/image_processing.py
@@ -18,6 +18,11 @@ import math
 from typing import Any, Optional, Tuple, Callable, Union, List, cast, Dict, TypeVar
 from numpy.typing import NDArray
 from typing import Any, Optional, Tuple, Callable, Union, List, TypeVar
+from skimage.restoration import denoise_bilateral
+from centrosome.filter import median_filter as _median_filter
+from centrosome.filter import circular_average_filter as _circular_average_filter
+from centrosome.smooth import fit_polynomial as _fit_polynomial
+from centrosome.smooth import smooth_with_function_and_mask as _smooth_with_function_and_mask
 from cellprofiler_library.types import ImageGrayscale, ImageGrayscaleMask, Image2DColor, Image2DGrayscale, Image2DGrayscaleMask, ImageAny, ImageAnyMask, ObjectSegmentation, Image2D, Image2DMask, StructuringElement, ObjectLabelSet, ImageColor, ImageBinaryMask
 from cellprofiler_library.opts import threshold as Threshold
 from cellprofiler_library.opts.enhanceorsuppressfeatures import SpeckleAccuracy, NeuriteMethod
@@ -28,8 +33,13 @@ from cellprofiler_library.opts.resize import ResizingMethod, DimensionMethod, In
 from cellprofiler_library.opts.imagemath import Operator
 from cellprofiler_library.opts.flipandrotate import RotationCoordinateAlignmnet
 from cellprofiler_library.opts.enhanceedges import EdgeDirection
+
 invert = cast(Callable[[ImageAny], ImageAny], _invert)
 isscalar = cast(Callable[[Optional[ImageAny]], bool], _isscalar)
+median_filter_centrosome = cast(Callable[[Image2D, Optional[Image2DMask], float], Image2D], _median_filter)
+circular_average_filter = cast(Callable[[Image2D, float, Optional[Image2DMask]], Image2D], _circular_average_filter)
+smooth_with_function_and_mask = cast(Callable[[Image2D, Callable[[Image2D], Image2D], Optional[Image2DMask]], Image2D], _smooth_with_function_and_mask)
+fit_polynomial = cast(Callable[[Image2D, Optional[Image2DMask], bool], Image2D], _fit_polynomial)
 
 T = TypeVar("T", bound=ImageAny)
 MorphImageT = TypeVar("Union[ImageGrayscale, ImageGrayscaleMask]", bound=Union[ImageGrayscale, ImageGrayscaleMask])
@@ -2042,3 +2052,44 @@ def rotate_image_coordinates(pixel_data: Image2D, mask: Image2DMask, rotate_poin
             "Unknown axis: %s" % rotate_coordinate_alignment.value
         )
     return angle
+
+
+###############################################################################
+# Smoothing
+###############################################################################
+
+def smoothing_gaussian(pixel_data: Image2D, mask: Optional[Image2DMask], sigma: float) -> Image2D:
+    def fn(image: Image2D) -> Image2D:
+            return scipy.ndimage.gaussian_filter(
+                image, sigma, mode="constant", cval=0
+            )
+    return smooth_with_function_and_mask(pixel_data, fn, mask)
+     
+
+def smoothing_median(pixel_data: Image2D, mask: Optional[Image2DMask], obj_size: float) -> Image2D:
+    return median_filter_centrosome(pixel_data, mask, obj_size / 2 + 1)
+
+
+def smoothing_keeping_edges(pixel_data: Image2D, multichannel: bool, sigma_range: float, sigma: float) -> Image2D:
+    assert sigma_range is not None, "sigma_range must be provided for smooth_keeping_edges"
+    return denoise_bilateral(
+        image=pixel_data.astype(float),
+        channel_axis=2 if multichannel else None,
+        sigma_color=sigma_range,
+        sigma_spatial=sigma,
+    )
+
+
+def smoothing_fit_polynomial(pixel_data: Image2D, mask: Optional[Image2DMask], clip: bool) -> Image2D:
+    return fit_polynomial(pixel_data, mask, clip)
+
+
+def smoothing_circular_average(pixel_data: Image2D, mask: Optional[Image2DMask], obj_size: float) -> Image2D:
+    return circular_average_filter(pixel_data, obj_size / 2 + 1, mask)
+
+def smoothing_smooth_to_average(pixel_data: Image2D, mask: Optional[Image2DMask]) -> Image2D:
+    if mask is not None:
+        mean = numpy.mean(pixel_data[mask])
+    else:
+        mean = numpy.mean(pixel_data)
+    return numpy.ones(pixel_data.shape, pixel_data.dtype) * mean
